Makefile: Remove SILENT and IGNORE
.IGNORE means that Make ignores errors.
This means that running `make test` exits with zero exit code even
if tests fail. This means it is no good for continuous integration,
Automated systems like Jenkins and Travis depend on exit codes to
detect failures.

.SILENT means that Make suppresses printing out what it will run.

If we want to ignore errors for some Make tasks we can prefix them
with a hyphen, see
https://www.gnu.org/software/make/manual/make.html#Errors

If we want to suppress printing particular commands, we can prefix
them with '@':
https://www.gnu.org/software/make/manual/make.html#Echoing