2/3: get file encoding from config file
Previously, get_dataframe_encoding was used to determine the
appropriate encoding to use for csv output using geopandas' to_csv.

This is difficult because a dataframe doesn't really have an encoding
of its own, so get_dataframe_encoding used chardet to try to guess
an encoding from the characters in the column names.

Under python 3, chardet refuses to do this because it will only
operate on byte streams, not strings. In order to convert the string
into a byte stream, you need an encoding.

I have worked around this by adding a configuration item for the
encoding, and reading that. If the item is not specified in the
config file, it defaults to utf-8, the default encoding used by
pandas.

utf-8 encoding should be fine for linux, mac, and windows 10. Older
versions of windows may need to specify utf-16.